+++ apps/boruta_gateway/lib/boruta_gateway/http_gateway.ex
@@ -339,7 +339,9 @@ defmodule BorutaGateway.HttpGateway do
             :ok =
               :gen_tcp.send(
                 client_socket,
-                transform_header(payload, upstream, token) <> request.body
+                transform_header(payload, upstream, token)
+                |> HttpRequest.put_header("X-Request-ID", request.request_id)
+                |> Kernel.<>(request.body)
               )
 
             upstream_start = :os.system_time(:microsecond)
@@ -391,7 +393,9 @@ defmodule BorutaGateway.HttpGateway do
             :ok =
               :ssl.send(
                 client_socket,
-                transform_header(payload, upstream, token) <> request.body
+                transform_header(payload, upstream, token)
+                |> HttpRequest.put_header("X-Request-ID", request.request_id)
+                |> Kernel.<>(request.body)
               )
 
             upstream_start = :os.system_time(:microsecond)